Cancel the death fade out when loading a replay
This would cause desyncs when loading a replay immediately after dying,
because once the fade out reaches 100% opacity the player is moved back
to the start of the level.

Note that this fix does not clear the FadeOutEvent, so checkpoints will
be loaded on the next fade out. As far as I can tell this is harmless.