Add camera visualisation
Draw the camera activation zones in dark blue, with lines where there
are links between multiple zones. Also draw each camera zone's
"transformed zone" in cyan (this is determined by a scaling and offset
from the activation zone).

If the player is inside a camera's activation zone, then camera's target
position is computed by transforming the player's current position by
the function that maps the camera's activation zone to its transformed
zone. If the player is inside the trapezium formed by two linked camera
zones, then the target camera position is a weighted linear
interpolation of the computed target positions of each camera zone.

Note that this visualisation does not indicate the zoom level, nor
strength (the weighting in the interpolation), or warmup/cooldown times
(how long the camera takes to get to full strength).