Fix replays creating multiple entries at time 0
When the game tries to load a replay it creates `Axis/ButtonReplay`
instances for each axis/button channel containing the inputs from the
replay that is being watched. The base class constructor means that
these are initialised with a "Release" entry at time 0. However, there
is also a "Release" entry at time 0 in the replay, so this results in
two entries existing at time 0, causing pain when saving the replay
later.

This patch fixes this bug by replacing the built-in replay loading
function with one that clears the initial entry that is created by the
base class constructor.