Show the reachable height from the current position and velocity
Due to the order that velocity and position are updated, simply moving
up and down loses energy, so the reachable height will fall. I have it
set up so that it will stay stable while moving upwards, and fall while
moving downwards. However, it seems to do whatever it likes when making
turns...