Use InputManager.OnFixedUpdate instead of Player.OnFixedUpdate
This fixes a specific cutscene in Home (the one where Quill gets the
cloak) in which the Player is not updated, but the InputManager is.

This may introduce off-by-one errors in places, because the InputManager
seems to always have a frame number one greater than the Player, due to
the order in which they are updated.