Do not skip over the first frame of replays
On the first frame of a replay, the AxisReplay.pointer was being set to
2 because the first timestamp that InputManager uses is 1, but the first
timestamp in replay events was 0. This meant that a single frame input
event at the start would be ignored. To fix this, I have added 1 to the
timestamp of every input event when creating the replay buffer.