Add QueueReplay command
This will append to a list of queued replays, and when the next scene is
loaded the first queued replay will be loaded. This lets us string
together multiple individual TAS files to create a fullgame TAS.

Unfortunately, there is no way to leave normal levels with just the
replay inputs once they have been completed, so for now manual
intervention is required to advance to the next level.