Revert time + 1 hack
After some more investigation, it turns out that in normal gameplay, no
inputs can be pressed on the very first frame. So none of the replays
that the TAS will be playing back should ever input an action on frame
1.

This means that the input event at time 0 will always be a release, so
the first InputManager update tick being at time 1 is no longer a
problem that we need to work around. Thus, the time + 1 hack can go.