Send the version string when loading replays
In older versions of the game, the physics were slightly different. So
in order to be able to watch replays from older versions of the game,
every replay contains a version number to tell the game which version it
was created in. This is used in the player script to change the physics.

In `bafc20c`, when making a small refactor to how replays are loaded, I
neglected to update the replay version, meaning that since that commit
every replay has been treated as if it had been created on the very
first version of the game!