Force the InputManager to be created in Player.Start
There are two different types of update methods in Unity:

1. FixedUpdate for physics stuff, which Unity will try to call a fixed
   number of times per second.
2. Update  for game logic that happens once per frame.

So one "frame" of the game will contain zero or more FixedUpdate calls
on all active game objects followed by one Update call for each game
object (see https://docs.unity3d.com/Manual/ExecutionOrder.html for a
timeline). In practise, this means that if the game is lagging, there
will be multiple FixedUpdate calls per Update call. The frame skip
option controls how many FixedUpdates are allowed before an Update is
required (up to a max of 3).

Initially, none of the scenes actually contain an InputManager. Instead,
when something wants it, they will call MasterController.GetInput(),
which will create an InputManager and add it to the scene if one does
not already exist.

Now, there are two places (that we care about) where this is called:

1. PauseMenu.Update
2. Player.FixedUpdate

Due to the nature of the two types of update methods, it is effectively
random which of these is called first:

- If PauseMenu.Update happens first, then the InputManager will be
  created during the Update step, meaning that it will exist when the
  next FixedUpdate step starts, so we get one Player.FixedUpdate and
  then an InputManager.FixedUpdate.
- On the other hand, if Player.FixedUpdate is called first, then the
  InputManager is only created after the FixedUpdate step has started,
  so doesn't get updated until the next frame, meaning that an extra
  Player.FixedUpdate happens before the first InputManager.FixedUpdate.

This patch fixes this non-determinism by forcing the InputManager to be
created in Player.Start, which will always happen before either of these
two cases, effectively giving us the "good luck" every time.