Keep a separate Frame counter and change OnPlayerUpdate to OnFrameEnd
This fixes a few bugs where certain objects in the scene were being
updated in a different order, causing some parts of the mod to be
delayed by a frame.

It also means that I don't need to worry (as much) about the
InputManager sometimes starting its frame count from 0 instead of 1.