Remove hack intended to prevent duplicate input entries
All this avoided was some warning logs when watching a replay. But it
had also introduced a bug by allowing entries at time 0 to be seen by
the player because I would always read entries up to time 0 when loading
replays. Now I only read entries when loading a savestate.